+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - nostr: remove `Market::Mention` (NIP-10) ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/895)
 - nostr: remove `parser` feature ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/899)
 - nostr: update `Nip19Profile::new` and `Nip19Coordinate::new` signature ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/910)
+- nostr: update `RelayInformationDocument::get` signature ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/913)
 - connect: remove `NostrConnect::get_relays` ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/894)
 
 ### Changed
@@ -41,6 +42,7 @@
 ### Added
 
 - nostr: add NIP-88 support ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/892)
+- nostr: add `Nip11GetOptions` ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/913)
 
 ### Fixed